fix 404 page theme in www
